Load multiple models to the framework

+++ reinkan/src/MemoryBinding/ReinkanVertexBuffer.cpp
@@ -42,7 +42,7 @@ namespace Reinkan
         appVertexBufferWrap = CreateStagedBufferWrap(vertices, VK_BUFFER_USAGE_VERTEX_BUFFER_BIT);
     }
 
-    void ReinkanApp::CreateIndexBuffer(std::vector<uint16_t> indices)
+    void ReinkanApp::CreateIndexBuffer(std::vector<uint32_t> indices)
     {
         appIndexBufferWrap = CreateStagedBufferWrap(indices, VK_BUFFER_USAGE_INDEX_BUFFER_BIT);
     }
